requirements.txt
# Core dependencies
jaseci>=1.4.0
jaseci-serv>=1.4.0

# Code parsing
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-java>=0.20.0

# Repository handling
gitpython>=3.1.0
requests>=2.31.0

# Documentation generation
markdown>=3.3.6,<3.4.0
graphviz>=0.20.0

# LLM integration
openai>=1.0.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

